add tank subclass with spring arm camera
this included a bit of blueprint hackery where we changed the base class
of BP_PawnTank from BasePawn to Tank

the build/compile order matters significantly. i ran into an issue when
altering the constructor where a placed instance of the blueprint did
not rerun the constructor. this may have been a specific issue with the
property assignment, but i can see this becoming an issue in the future.

see: https://www.gamedev.tv/courses/1638644/lectures/39269599