delay start and add crosshair
- custom controller class to encapsulate input/crosshair toggle
- add delayed start logic to game mode class
- remove existing debugsphere crosshair placeholder

this lecture was helpful in that it surfaced another delegate pattern in
unreal that allows binding paramaters to be passed to an asynchronous
invocation. this didn't at all cover async behaviour or talk about state
or scope in which the execution time of one of these bound functions may
read values, but i expect to see the pattern used in other places in
unreal outside of the timer example provided.

see: https://www.gamedev.tv/courses/1638644/lectures/39276647
see: https://www.gamedev.tv/courses/1638644/lectures/39276647